// apps/web/src/dialogs/settings/types.ts
const SectionKeys = [
  "profile",
  "auth",
  "subscription",
  "sync",
  "appearance",
  "behaviour",
  "desktop",
  "notifications",
  "servers",
  "editor",
  "backup-export",
  "export",
  "importer",
  "vault",
  "app-lock",
  "privacy",
  "support",
  "legal",
  "developer",
  "about",
  "inbox",
  "circle"
] as const;

export type SectionKeys = (typeof SectionKeys)[number];

export type SectionGroupKeys =
  | "account"
  | "customization"
  | "import-export"
  | "security"
  | "other";

export type SectionGroup = {
  key: SectionGroupKeys;
  title: string;
  sections: Section[];
};

export type Section = {
  key: SectionKeys;
  title: string;
  icon: Icon;
  isHidden?: () => boolean;
};

export type SettingsGroup = {
  key: string;
  section: SectionKeys;
  settings: Setting[];
  header: string | ((props: any) => JSX.Element | null);
  isHidden?: () => boolean;
  onStateChange?: (
    listener: (state: any, prevState: unknown) => void
  ) => () => void;
  onRender?: () => void | Promise<void>;
};

export type Setting = {
  key: string;
  keywords?: string[];
  title: string;
  description?: string | ((state?: any) => string);
  components: SettingComponent[] | ((state?: any) => SettingComponent[]);
  isHidden?: (state?: any) => boolean;
  onStateChange?: (
    listener: (state: unknown, prevState: unknown) => void
  ) => () => void;
  featureId?: FeatureId;
};

export type SettingComponentType =
  | "toggle"
  | "dropdown"
  | "button"
  | "input"
  | "icon"
  | "custom";

export type SettingComponent =
  | ButtonSettingComponent
  | ToggleSettingComponent
  | IconSettingComponent
  | DropdownSettingComponent
  | InputSettingComponent
  | CustomSettingComponent;